bug1074906 enable screen reader in accessibilty settings
