+++ apps/settings/elements/accessibility.html
@@ -13,7 +13,7 @@
             <small id="colorFilter-desc">Disabled</small>
           </a>
         </li>
-        <li data-show-name="accessibility.screenreader-show-settings" hidden>
+        <li>
           <a class="menu-item" href="#accessibility-screenreader">
             <span data-l10n-id="screenReader">Screen Reader</span>
             <small id="screenReader-desc">Disabled</small>
