+++ apps/settings/elements/accessibility_screenreader.html
@@ -6,7 +6,7 @@
     </gaia-header>
 
     <div>
-      <ul data-show-name="accessibility.screenreader-show-settings" hidden>
+      <ul>
         <li id="screenreader-enable" class="hint">
           <label class="pack-switch">
             <input type="checkbox" data-ignore>
