+++ apps/settings/elements/developer.html
@@ -207,12 +207,6 @@
         </li>
         <li>
           <label class="pack-checkbox">
-            <input type="checkbox" name="accessibility.screenreader-show-settings">
-            <span data-l10n-id="showScreenReaderSettings">Show screen reader settings</span>
-          </label>
-        </li>
-        <li>
-          <label class="pack-checkbox">
             <input type="checkbox" name="dom.mozApps.use_reviewer_certs">
             <span data-l10n-id="use-reviewer-certs">Use Marketplace reviewer certs</span>
           </label>
