+++ tests/python/gaia-ui-tests/gaiatest/apps/settings/regions/accessibility.py
@@ -27,12 +27,18 @@ class Accessibility(Base):
 
 class AccessibilityScreenreader(Base):
 
+    _screen_reader_enable_switch_locator = (
+      By.CSS_SELECTOR, '#screenreader-enable input[type="checkbox"]')
     _screen_reader_captions_locator = (
         By.CSS_SELECTOR, 'input[name="accessibility.screenreader-captions"]')
     _screen_reader_volume_slider_locator = (
         By.CSS_SELECTOR, 'input[name="accessibility.screenreader-volume"]')
     _screen_reader_rate_slider_locator = (
         By.CSS_SELECTOR, 'input[name="accessibility.screenreader-rate"]')
+
+    def a11y_toggle_screen_reader(self):
+        self.accessibility.click(self.marionette.find_element(
+            *self._screen_reader_enable_switch_locator))
 
     def a11y_toggle_captions(self):
         self.accessibility.click(self.marionette.find_element(
