+++ tests/python/gaia-ui-tests/gaiatest/tests/accessibility/settings/manifest.ini
@@ -4,3 +4,4 @@ b2g = true
 [test_a11y_slider_visibility.py]
 [test_a11y_color_filters.py]
 [test_a11y_captions.py]
+[test_a11y_screen_reader.py]
