+++ tests/python/gaia-ui-tests/gaiatest/tests/accessibility/settings/test_a11y_color_filters.py
@@ -10,7 +10,7 @@ class TestColorFiltersAccessibility(GaiaTestCase):
     def setUp(self):
         GaiaTestCase.setUp(self)
 
-        # make accessibility settings visible
+        # set initial state, including colors disabled
         self.data_layer.set_setting('accessibility.colors.enabled', False)
         self.data_layer.set_setting('accessibility.colors.invert', False)
         self.data_layer.set_setting('accessibility.colors.grayscale', False)
